# Conflicts:
#	app/build.gradle
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/ch/stephgit/memory/MainActivity.kt
#	app/src/main/java/ch/stephgit/memory/MemoryApp.kt
#	app/src/main/java/ch/stephgit/memory/UserFragment.kt
#	app/src/main/java/ch/stephgit/memory/onboarding/LoginFragment.kt
#	app/src/main/java/ch/stephgit/memory/onboarding/RegistrationFragment.kt
#	app/src/main/java/ch/stephgit/memory/persistence/AppDatabase.kt
#	app/src/main/java/ch/stephgit/memory/persistence/entity/Player.kt
#	app/src/main/java/ch/stephgit/memory/persistence/repository/PlayerRepository.kt
#	app/src/main/java/ch/stephgit/memory/ui/main/HistoryFragment.kt
#	app/src/main/java/ch/stephgit/memory/ui/onboarding/OnboardingActivity.kt
#	app/src/main/java/ch/stephgit/memory/ui/onboarding/ProfileFragment.kt

+++ app/src/main/java/ch/stephgit/memory/GamePlayFlow.kt
@@ -2,5 +2,5 @@ package ch.stephgit.memory
 
 interface GamePlayFlow {
 
-    fun goToResult(flips: Int)
+    fun goToResult(flips: Long)
 }
